+++ include/threepp/renderers/gl/ProgramParameters.hpp
@@ -129,7 +129,7 @@ namespace threepp {
             ProgramParameters(
                     const GLRenderer &renderer,
                     const GLLights::LightState &lights,
-                    int numShadows,
+                    size_t numShadows,
                     Object3D* object,
                     Scene* scene,
                     Material *material,
