+++ src/threepp/renderers/gl/GLPrograms.cpp
@@ -45,7 +45,7 @@ ProgramParameters GLPrograms::getParameters(
         const GLRenderer &renderer,
         Material *material,
         const GLLights::LightState &lights,
-        int numShadows,
+        size_t numShadows,
         Scene* scene,
         Object3D* object) {
